Added Docker Dev Container (#2)
* Able to get Docker in Docker with devcontainer config

This commit adds some base ROS docker files for the project along with modifications to devcontainer.json and Dockerfile to allow the use of docker inside of the dev Docker container
This currently means that the docker container is able to use the host/local docker daemon from outside the container to build images. Currently requires the use of sudo, will
need to see if we can remove that requirement somehow. Maybe mapping the user directly inside the container?

* Some more modifications to Docker in Docker setup, can run catkin_make inside the dev container as well