Implemented bulk of A* search
this commit (done late at night) implements the bulk of the A* search for the planner. At this stage,
the code is able to compile and thats it

This commit made substantial changes to various data structures: operator overloading for Waypoint, addition
of equal_to for Node, changed constructor of Node to pass by value instead of reference of the from_
member variable initialization. There are probably high penalities for that. If not not noted already,
some of the changes were largely for using unordered_set and shared pointers for PQ.