Add suas_planning pacakge's A* implementation (#8)
* Address hashing of Node for use in unordered set

This commit adds custom  hashing implementation for nodes to allow for use of unorderset as an priority queue

* Finally actually compiled source files and fixed errors

This commit makes changes for the successful compilation of code. Now that the node is setup with the
Cmake file configured correctly, we actually are doing work and have things to fix

Lessons learned: It still C++ and its still ancient so order of things matter and the author of this
commit should actually go read the Effective C++ book he has and actually properly learn it for once

* Started implementation of A* search

This commit is just a placeholder to commit current start of modiciations to implement A*
Note that is largely not done yet and is to serve as an marker for other signficant changes like the use
of a shared_ptr in the Node class and changes to the priority queue as such. Also changes the comprator
from using std::greater to a lambda function to dereference the pointers first before comparison

* Implemented bulk of A* search

this commit (done late at night) implements the bulk of the A* search for the planner. At this stage,
the code is able to compile and thats it

This commit made substantial changes to various data structures: operator overloading for Waypoint, addition
of equal_to for Node, changed constructor of Node to pass by value instead of reference of the from_
member variable initialization. There are probably high penalties for that. If not not noted already,
some of the changes were largely for using unordered_set and shared pointers for PQ.

* Added a minimal launch file

* Updated some names

* changed iterator for loop to ranged for loop syntax

* Updated all C style type casts to safer static_cast where appropriate

* Change compute map functions to use waypoints and min/max_element from stl

* Fixed typos